Reorganize functions in resolvers.ts
