+++ server/graphql/utils/requestGroup.ts
@@ -1,5 +1,28 @@
 import mongoose, { Document } from 'mongoose'
-import { softDeleteRequestTypeHelper } from './requestType'
+
+import { nextRequestRequestTypeHelper, softDeleteRequestTypeHelper } from './requestType'
+import { RequestInterface } from '../../models/requestModel'
+
+const nextRequestRequestGroupHelper = (requestTypeIds, dataSources): RequestInterface => {
+  const requests = requestTypeIds.map((id) => {
+    const requestType = dataSources.requestTypes.getById(id)
+    return nextRequestRequestTypeHelper(requestType.requests, dataSources)
+  })
+  requests.sort((request1, request2) => {
+    if (!request1 && !request2) {
+      return 0
+    }
+    if (!request1) {
+      return 1
+    }
+    if (!request2) {
+      return -1
+    }
+    return request1.dateCreated - request2.dateCreated
+  })
+  return requests.length == 0 ? null : requests[0]
+}
+
 
 const updateRequestGroupHelper = async (requestGroup, dataSources, session): Promise<Document> => {
   const res = await dataSources.requestGroups.update(requestGroup, session)
@@ -28,4 +51,4 @@ const softDeleteRequestGroupHelper = async (id, dataSources) => {
   }
 }
 
-export { softDeleteRequestGroupHelper, updateRequestGroupHelper }
+export { nextRequestRequestGroupHelper, softDeleteRequestGroupHelper, updateRequestGroupHelper }
